Refactor tests for data ingestion and add UAT notebook
Reorganized and expanded unit tests for `ImageDataIngestorImpl` to include additional validation checks. Added a new user acceptance testing (UAT) notebook for dataset ingestion and image display validation. Updated `requirements.txt` to include `pillow` and `scikit-learn` for the new functionality.
